Decouple compiler from alloy
