+++ packages/compiler/src/core/program.ts
@@ -1,4 +1,3 @@
-import { type OutputDirectory } from "@alloy-js/core";
 import { EmitterOptions } from "../config/types.js";
 import { createAssetEmitter } from "../emitter-framework/asset-emitter.js";
 import { setCurrentProgram } from "../experimental/typekit/index.js";
@@ -18,7 +17,6 @@ import { createBinder } from "./binder.js";
 import { Checker, createChecker } from "./checker.js";
 import { createSuppressCodeFix } from "./compiler-code-fixes/suppress.codefix.js";
 import { compilerAssert } from "./diagnostics.js";
-import { emitFile } from "./emitter-utils.js";
 import { resolveTypeSpecEntrypoint } from "./entrypoint-resolution.js";
 import { ExternalError } from "./external-error.js";
 import { getLibraryUrlsLoaded } from "./library.js";
@@ -578,28 +576,9 @@ export async function compile(
       },
     };
     try {
-      const result = (await emitter.emitFunction(context)) as any;
-      if (typeof result === "function") {
-        const { render } = await import("@alloy-js/core");
-        // assume this is an alloy component
-        const tree = render(result);
-        await writeOutputDirectory(tree, context.emitterOutputDir);
-      }
+      await emitter.emitFunction(context)
     } catch (error: unknown) {
       throw new ExternalError({ kind: "emitter", metadata: emitter.metadata, error });
-    }
-  }
-
-  async function writeOutputDirectory(dir: OutputDirectory, emitterOutputDir: string) {
-    for (const sub of dir.contents) {
-      if (Array.isArray(sub.contents)) {
-        await writeOutputDirectory(sub as OutputDirectory, emitterOutputDir);
-      } else {
-        await emitFile(program, {
-          content: sub.contents as string,
-          path: joinPaths(emitterOutputDir, sub.path),
-        });
-      }
     }
   }
