+++ packages/emitter-framework/src/core/index.ts
@@ -1,2 +1,3 @@
 export * from "./context/index.js";
 export * from "./transport-name-policy.js";
+export * from "./write-output.js";
